+++ sdk/eventgrid/arm-eventgrid/samples-dev/topicTypesListEventTypesSample.ts
@@ -1,25 +1,23 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-/**
- * This sample demonstrates how to List event types for a topic type.
- *
- * @summary List event types for a topic type.
- * x-ms-original-file: specification/eventgrid/resource-manager/Microsoft.EventGrid/preview/2025-04-01-preview/examples/TopicTypes_ListEventTypes.json
- */
-
 import { EventGridManagementClient } from "@azure/arm-eventgrid";
 import { DefaultAzureCredential } from "@azure/identity";
-import "dotenv/config";
 
+/**
+ * This sample demonstrates how to list event types for a topic type.
+ *
+ * @summary list event types for a topic type.
+ * x-ms-original-file: 2025-07-15-preview/TopicTypes_ListEventTypes.json
+ */
 async function topicTypesListEventTypes(): Promise<void> {
-  const topicTypeName = "Microsoft.Storage.StorageAccounts";
   const credential = new DefaultAzureCredential();
   const client = new EventGridManagementClient(credential);
   const resArray = new Array();
-  for await (const item of client.topicTypes.listEventTypes(topicTypeName)) {
+  for await (const item of client.topicTypes.listEventTypes("Microsoft.Storage.StorageAccounts")) {
     resArray.push(item);
   }
+
   console.log(resArray);
 }