+++ sdk/eventgrid/arm-eventgrid/src/api/partnerRegistrations/index.ts
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export {
+  listBySubscription,
+  listByResourceGroup,
+  $delete,
+  update,
+  createOrUpdate,
+  get,
+} from "./operations.js";
+export type {
+  PartnerRegistrationsListBySubscriptionOptionalParams,
+  PartnerRegistrationsListByResourceGroupOptionalParams,
+  PartnerRegistrationsDeleteOptionalParams,
+  PartnerRegistrationsUpdateOptionalParams,
+  PartnerRegistrationsCreateOrUpdateOptionalParams,
+  PartnerRegistrationsGetOptionalParams,
+} from "./options.js";