+++ sdk/kubernetesconfiguration/arm-kubernetesconfiguration-fluxconfigurations/src/api/fluxConfigurations/options.ts
@@ -1,0 +1,30 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import type { OperationOptions } from "@azure-rest/core-client";
+
+/** Optional parameters. */
+export interface FluxConfigurationsListOptionalParams extends OperationOptions {}
+
+/** Optional parameters. */
+export interface FluxConfigurationsDeleteOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** Delete the extension resource in Azure - not the normal asynchronous delete. */
+  forceDelete?: boolean;
+}
+
+/** Optional parameters. */
+export interface FluxConfigurationsUpdateOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+}
+
+/** Optional parameters. */
+export interface FluxConfigurationsCreateOrUpdateOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+}
+
+/** Optional parameters. */
+export interface FluxConfigurationsGetOptionalParams extends OperationOptions {}