+++ sdk/web-pubsub/web-pubsub-client/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 1.0.4 (2026-03-15)
+## 1.0.5 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.4 (2026-03-20)
 
 ### Features Added